% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{bubs}
\alias{bubs}
\title{Make a bubble plot to look at residuals}
\usage{
bubs(x, y, z, xlim = c(0, 55), ylim = c(1969, 2016),
  xlab = "Length (cm)", ylab = "Year")
}
\arguments{
\item{x, y, z}{the independent, dependent and residual variables}
}
\description{
Make a bubble plot to look at residuals
}
