% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{confint.f}
\alias{confint.f}
\title{Draw polygon confidence intervals}
\usage{
confint.f(x, ylow, yhigh, ...)
}
\description{
Draw polygon confidence intervals
}
\examples{

}
\keyword{function}
\keyword{helper}
