% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.object.helper.functions.r
\name{confint}
\alias{confint}
\title{Draw polygon confidence intervals}
\usage{
confint(x, ylow, yhigh, ...)
}
\arguments{
\item{x}{the independent variable vector}

\item{ylow}{the lower confidence interval y series}

\item{yhigh}{the higher confidence interval y series}

\item{...}{other argument from par that can be used with the polygon function}
}
\description{
just a call to polygon but with more intuitive arguments in the context on making confidence intervals.
     You need to call this before you overlay your median with the line function.
}
\examples{

}
\keyword{function}
\keyword{helper}
