% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.helper.functions.r
\name{count.distinct}
\alias{count.distinct}
\title{Counts distinct values of a vector}
\usage{
count.distinct(x)
}
\arguments{
\item{x}{The vector for which you want distinct values counted}
}
\value{
a scalar with the count of unique values in a vector
}
\description{
mimics count(distinct value) in mysql when using a group by statement. Used most in apply.
}
\examples{
mat=matrix(rep(1:10,each=10),ncol=10)
apply(mat,2,length); apply(mat,2,count.distinct)
apply(mat,1,length); apply(mat,1,count.distinct)
}
\author{
Daniel Duplisea
}
