% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.helper.functions.r
\name{decdeg.f}
\alias{decdeg.f}
\title{Convert geoegraphic coordinates in degrees minutes seconds to decimal degrees}
\usage{
decdeg.f(degminsec)
}
\arguments{
\item{degminsec}{all in one number without commas or decimals}
}
\description{
Convert geoegraphic coordinates in degrees minutes seconds to decimal degrees
}
\author{
Daniel Duplisea
}
