% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.helper.functions.r
\name{diags.f}
\alias{diags.f}
\title{Extract sub-diagonals of a matrix}
\usage{
diags.f(mat, diag.pos = -1)
}
\arguments{
\item{mat}{data matrix}

\item{diag.pos}{the sub-diagonal position, i.e. -1 extract the first lower subdiagonal, 0 is the main diagonal}
}
\value{
the specified diagonal as vector
}
\description{
I have only used it for Leslie matrix modelling. Not sure where else it would be useful.
}
\examples{
data.matrix= matrix(1:100,ncol=10)
diags.f(data.matrix,0)
diags.f(data.matrix,-2)
diags.f(data.matrix,8)
}
\author{
Daniel Duplisea
}
