% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{error.bars.data.f}
\alias{error.bars.data.f}
\title{Put error bars on an x y plot given high and low data series}
\usage{
error.bars.data.f(x, error.low, error.high, ...)
}
\arguments{
\item{x}{independent data series}

\item{error.low}{the time series of lower values}

\item{error.high}{the time series of upper values}

\item{SE}{if T then y.sd is the standard error}
}
\description{
Put error bars on an x y plot given high and low data series
}
\examples{

}
