% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{error.bars.sd.f}
\alias{error.bars.sd.f}
\title{Put error bars on an x y plot given a standard deviation of normal distributed errors}
\usage{
error.bars.sd.f(x, y.mean, y.sd, SE = T, ...)
}
\arguments{
\item{x}{independent data series}

\item{y.mean}{mean or central value of the y series}

\item{y.sd}{the standard deviation of the y series}

\item{SE}{if T then y.sd is the standard error}
}
\description{
Put error bars on an x y plot given a standard deviation of normal distributed errors
}
\examples{

}
