% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{gam.trend.line.f}
\alias{gam.trend.line.f}
\title{Put a gam trendline on a plot}
\usage{
gam.trend.line.f(x, y, ...)
}
\arguments{
\item{vec}{vector of data}
}
\value{
pulls/pops off the final value of a vector4
}
\description{
pops off the final value on a vector. You do not need to assign the vector to itself again, it is done internally so
        somewhat against the R way of doing things and it could get you into trouble.
}
\author{
Daniel Duplisea
}
