% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{getBathy}
\alias{getBathy}
\title{Plot bathymetry on PBSmapping map from a data object}
\usage{
getBathy(bathyFname, isob = c(100, 200, 500, 800, 1200, 2400),
  minVerts = 3)
}
\description{
The default dataset is for eastern Canada. You need to go to NOAA to get data for your region.
        You will likely need to use the joinBathy to make your downloaded data work properly
}
\examples{

}
\references{
http://maps.ngdc.noaa.gov/viewers/wcs-client/
}
\keyword{function}
\keyword{helper}
