% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.object.helper.functions.r
\name{lmP.f}
\alias{lmP.f}
\title{functions to extract P fit diagnostics from an lm object}
\usage{
lmP.f(lm.fit.object)
}
\arguments{
\item{lm.fit.object}{a linear model fit objects}
}
\value{
The p value
}
\description{
Calculated the p value from a linear model fit (ANOVA F-test)
}
\examples{
fake.data= data.frame(x=1:100, y=.05*(1:100)^2 - 5*(1:100) + 2 + rnorm(100,0,5)) #fit a linear to quadratic just for the hell of it
plot(fake.data)
lm.fit.object= lm(y~ x, data=fake.data)
lines(1:100,predict(lm.fit.object))
lmP.f(lm.fit.object=lm.fit.object)
}
