% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{map.f}
\alias{map.f}
\title{Wrapper for the PBSmapping plotMap function}
\usage{
map.f(longs = c(-74, -50), lats = c(43, 52.25),
  land.colour = "sienna3", sea.colour = "lightblue", bathy = T,
  isob = c(100, 200))
}
\arguments{
\item{longs}{longitude decimal degrees (- for western hemisphere)}

\item{lats}{latitude decimal degrees (- for southern hemisphere)}

\item{bathy}{do you want to show bathymetric lines. Default = T}

\item{isob}{the depth of the isobaths you want to show}

\item{land.col}{colour of the land}

\item{sea.col}{colour of the oceans and water bodies}
}
\description{
Make a map using PBSmapping
}
\examples{

}
