% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.helper.functions.r
\name{map.inset.f}
\alias{map.inset.f}
\title{An inset map for the main map function}
\usage{
map.inset.f(longs = c(-71, -50), lats = c(43, 52.25),
  land.colour = "sienna3", sea.colour = "lightblue")
}
\arguments{
\item{longs}{longitude decimal degrees (- for western hemisphere)}

\item{lats}{latitude decimal degrees (- for southern hemisphere)}

\item{land.col}{colour of the land}

\item{sea.col}{colour of the oceans and water bodies}
}
\description{
A simple stripped down map without annotation that can be used as an inset to the larger scale map
}
\examples{

}
\keyword{function}
\keyword{helper}
