% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.helper.functions.r
\name{pop}
\alias{pop}
\title{Pop the final value off the end of vector}
\usage{
pop(vec)
}
\arguments{
\item{vec}{vector of data}
}
\value{
pulls/pops off the final value of a vector4
}
\description{
pops off the final value on a vector. You do not need to assign the vector to itself again, it is done internally so
        somewhat against the R way of doing things and it could get you into trouble.
}
\references{
Kevin Wright http://www.johnmyleswhite.com/notebook/2009/12/07/implementing-push-and-pop-in-r/#comment-17810
}
\author{
Kevin Wright
}
