% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.helper.functions.r
\name{push}
\alias{push}
\title{Push a value onto the end of vector}
\usage{
push(vec, item)
}
\arguments{
\item{vec}{vector of data}

\item{item}{the new values you want to append}
}
\value{
a vector with new values pushed onto the end of the old vector
}
\description{
pushes values onto the end of vector. Used by data science people when they do not know how big their
        final object size should be. You do not need to assign the vector to itself again, it is done internally so
        somewhat against the R way of doing things and it could get you into trouble
}
\references{
Kevin Wright http://www.johnmyleswhite.com/notebook/2009/12/07/implementing-push-and-pop-in-r/#comment-17810
}
\author{
Kevin Wright
}
