% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.helper.functions.r
\name{decdeg.f}
\alias{decdeg.f}
\title{Convert geoegraphic coordinates in degrees minutes seconds to decimal degrees}
\usage{
decdeg.f(degminsec)
}
\arguments{
\item{degminsec}{all in one number without commas or decimals, e.g. 44 38' 52", input as 443852}
}
\description{
a function to convert degrees, minutes, seconds coordinates into decimal degrees.
}
\author{
Daniel Duplisea
}
